Deduplicate ConnectorManifest into shared crate
