+++ connectors/slack/src/models.rs
@@ -9,23 +9,7 @@ use time::OffsetDateTime;
 // Connector Protocol Models
 // ============================================================================
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConnectorManifest {
-    pub name: String,
-    pub version: String,
-    pub sync_modes: Vec<String>,
-    #[serde(default)]
-    pub actions: Vec<ActionDefinition>,
-    #[serde(default)]
-    pub search_operators: Vec<shared::models::SearchOperator>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ActionDefinition {
-    pub name: String,
-    pub description: String,
-    pub parameters: JsonValue,
-}
+pub use shared::models::{ActionDefinition, ConnectorManifest};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct SyncResponse {
